change achieved data method from DB to JSON

+++ src/main/java/entity/Category.java
@@ -13,5 +13,5 @@ public class Category {
     private Integer categoryId;
     private String categoryName;
     private Double price;
-    private Float priceMultiple;
+    private Double priceMultiple;
 }
